Use serverless oauth function
This change uses standard Github authentication via a custom
authentication server running on AWS Lambda.

The engine uses the https://github.com/kpopper/netlify-serverless-oauth2-backend
code and is deployed to AWS Lambda via the Serverless framework.

The `prod` is a reference to the `prod` stage this is deployed to in
Lambda