Sort out SASS mess
We were previously using a SASS plugin to manage the SCSS>CSS translation. This is now built into Jekyll so switching to use the built-in approach.

Also removing normalize as it was causing jekyll build some problems. It looks like a newer version of normalize is being auto-generated into a file called `assets/css/style.css`. When I include this in the page, the layout goes weird but it may well be worth exploring an update to my styles to accommodate this.

Also worth exploring removing all the bourbon / neat stuff which is old now. Perhaps Tailwind would be better?